# backend/API_DOCS.yaml
openapi: 3.0.0
info:
  title: RAG Learning Path API
  description: API cho hệ thống tạo lộ trình học tập cá nhân hóa
  version: 1.1.0
  contact:
    name: PersonalEDU Support
    email: support@personaledu.example.com
servers:
  - url: http://localhost:5000
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Thông báo lỗi
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID người dùng
        username:
          type: string
          description: Tên đăng nhập
        email:
          type: string
          format: email
          description: Email người dùng
        created_at:
          type: string
          format: date-time
          description: Thời điểm tạo tài khoản
      required:
        - id
        - username
        - email
        - created_at
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: Token xác thực
      required:
        - user
        - token
...
paths:
  /api/auth/register:
    post:
      summary: Đăng ký tài khoản mới
      description: Tạo tài khoản người dùng mới và trả về token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Tên đăng nhập
                email:
                  type: string
                  format: email
                  description: Email
                password:
                  type: string
                  format: password
                  description: Mật khẩu
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Lỗi server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      summary: Đăng nhập
      description: Đăng nhập và lấy token xác thực
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Tên đăng nhập
                password:
                  type: string
                  format: password
                  description: Mật khẩu
              required:
                - username
                - password
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Thông tin đăng nhập không chính xác
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Lỗi server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      summary: Đăng xuất
      description: Đăng xuất và hủy token hiện tại
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đăng xuất thành công
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      summary: Thông tin người dùng hiện tại
      description: Lấy thông tin người dùng từ token
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thông tin người dùng
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/users/{username}/exists:
    get:
      summary: Kiểm tra tên người dùng
      description: Kiểm tra xem tên người dùng đã tồn tại chưa
      tags:
        - Authentication
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: Tên người dùng cần kiểm tra
      responses:
        '200':
          description: Kết quả kiểm tra
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                    description: Tên người dùng đã tồn tại hay chưa
...